% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadWrite.R
\name{writeMeshFile}
\alias{writeMeshFile}
\title{Export mesh to a file}
\usage{
writeMeshFile(mesh, filename, precision = 17L, binary = FALSE)
}
\arguments{
\item{mesh}{a mesh given either as a list containing (at least) the fields
\code{vertices} and \code{faces}, otherwise a \strong{rgl} mesh
(i.e. a \code{mesh3d} object)}

\item{filename}{name of the file to be written, with extension \code{stl},
\code{ply}, \code{obj} or \code{off}}

\item{precision}{positive integer, number of decimal digits for the vertices}

\item{binary}{Boolean, whether to write a binary file or an ASCII file}
}
\value{
No value, just generates the file.
}
\description{
Export a mesh to a file.
}
